requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
moviepy>=1.0.3
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
aiofiles>=23.2.0
